split model and generator
